adjust/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
appdirs==1.4.4
audioread==2.1.9
cachetools==4.2.2; python_version ~= '3.5'
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4; python_version >= '3'
decorator==5.1.0; python_version >= '3.5'
google-api-core==2.0.1; python_version >= '3.6'
google-auth==2.0.2; python_version >= '3.6'
google-cloud-core==2.0.0; python_version >= '3.6'
google-cloud-storage==1.42.1
google-crc32c==1.1.2; python_version >= '3.6'
google-resumable-media==2.0.2; python_version >= '3.6'
googleapis-common-protos==1.53.0; python_version >= '3.6'
idna==3.2; python_version >= '3'
joblib==1.0.1; python_version >= '3.6'
librosa==0.8.1
llvmlite==0.37.0; python_version < '3.10' and python_version >= '3.7'
numba==0.54.0; python_version < '3.10' and python_version >= '3.7'
numpy==1.20.3
packaging==21.0; python_version >= '3.6'
pooch==1.5.1; python_version >= '3.6'
protobuf==3.17.3
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.20; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyparsing==2.4.7; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
requests==2.26.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
resampy==0.2.2
rsa==4.7.2; python_version >= '3.5' and python_version < '4'
scikit-learn==0.24.2; python_version >= '3.6'
scipy==1.7.1
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
soundfile==0.10.3.post1
threadpoolctl==2.2.0; python_version >= '3.6'
urllib3==1.26.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
